update server btn
작업 중(노트북 업로드용)

+++ demo_battle/demo_battle/Player.cpp
@@ -130,7 +130,7 @@ void Player::SetCoin(size_t coin) {
 	m_coin_cnt = coin;
 }
 
-size_t Player::GetCoin() {
+int Player::GetCoin() {
 	return m_coin_cnt;
 }
